requirements.txt
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
Flask==1.1.1
requests==2.22.0
